# Conflicts:
#	nominate_app/templates/nominate_app/edit_award_template.html
#	nominate_app/views/award_template_views.py

+++ nominate_app/views/award_template_views.py
@@ -10,29 +10,32 @@ import pdb
 
 def new_award_template(request, award_id):
   award = Awards.objects.get(id = award_id)
+  is_active=False
   AwardTemplateFormset = modelformset_factory(Questions, fields=('qname', 'qtype', 'role', 'attachment_need'), extra=1, can_delete=True)
   if request.method == 'POST':
     new_form = request.POST.copy()
     template_name = new_form.pop('template_name')[0] 
-    is_active_val = new_form.pop('is_active')[0] 
-    is_active = True if is_active_val ==  'yes' else False
+    is_active_val = new_form.pop('is_active', ['off'])
+    is_active = True if is_active_val[0] ==  'on' else False
     award_id = award.id
-    award_template = AwardTemplate.objects.create(template_name= template_name, is_active=is_active, award_id=award_id)
-    request.POST = new_form
-    formset = AwardTemplateFormset(request.POST)
-    if formset.is_valid():
-      instances = formset.save(commit=False)
-      for obj in formset.deleted_objects:
-        obj.delete()
-      for instance in instances:
-        instance.award_template_id=award_template.id
-        instance.save()
-      return redirect('nominate_app:award_template_index')
-
+    if (template_name != ''):
+      award_template =AwardTemplate.objects.create(template_name= template_name, is_active=is_active, award_id=award_id)
+      request.POST = new_form
+      formset = AwardTemplateFormset(request.POST)
+      if formset.is_valid():
+        instances = formset.save(commit=False)
+        for obj in formset.deleted_objects:
+          obj.delete()
+        for instance in instances:
+          instance.award_template_id=award_template.id
+          instance.save()
+        return redirect('nominate_app:award_template_index')
+    else:
+      return render(request, 'nominate_app/new_award_template.html', {'formset':formset, 'template_name': award.name, 'form_errors': 'Template name is required'})
   else:
     formset = AwardTemplateFormset(queryset=Questions.objects.none())
 
-  return render(request, 'nominate_app/new_award_template.html', {'formset':formset, 'template_name': award.name})
+  return render(request, 'nominate_app/new_award_template.html', {'formset':formset, 'template_name': award.name, 'is_active':is_active})
 
 def edit_award_template(request, template_id):
   award_template = AwardTemplate.objects.get(id = template_id)
@@ -44,8 +47,11 @@ def edit_award_template(request, template_id):
     x=1
   AwardTemplateFormset = modelformset_factory(Questions, fields=('qname', 'qtype', 'role', 'attachment_need'), extra=x, can_delete=True)
   if request.method == 'POST':
+    is_active_val = request.POST.get('is_active',False)
+    is_active = True if is_active_val ==  'on' else False
     formset = AwardTemplateFormset(request.POST, queryset=questions)
     if formset.is_valid():
+      AwardTemplate.objects.filter(id=template_id).update(is_active=is_active)
       instances = formset.save(commit=False)
       for obj in formset.deleted_objects:
         obj.delete()
@@ -57,7 +63,7 @@ def edit_award_template(request, template_id):
   else:
     formset = AwardTemplateFormset(queryset=questions)
 
-  return render(request, 'nominate_app/edit_award_template.html', {'formset':formset, 'award_template': award_template})
+  return render(request, 'nominate_app/edit_award_template.html', {'formset':formset, 'award_template': award_template, 'is_active':award_template.is_active })
 
 def delete_award_template(request, ques_id):
   questions = Questions.objects.filter(id=ques_id)
